chore: integration, e2e 파일 구조 개선
- test/ 로 동시성 테스트 이동
- test/integration, test/e2e 테스팅 구성
- package.json script 추가
