+++ README.md
@@ -2,3 +2,12 @@
 
 - `./docs/concurrency.md` - 동시성 처리 리서치
 - `./docs/points-policy.md` - 정책 정리
+
+## 테스팅 방법
+
+```bash
+pnpm i
+npm run test
+npm run test:e2e
+npm run test:integration
+```
